led/requirements.txt
paho-mqtt==1.3.1
RPi.GPIO==0.6.3
